% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateInstruments.R
\name{candidateInstruments}
\alias{candidateInstruments}
\title{Extract candidate trait instruments}
\usage{
candidateInstruments(dat, include_outliers = FALSE)
}
\arguments{
\item{dat}{Output from outlierAssociations() function}

\item{include_outliers}{When performing MR of candidate traits against exposures or outcomes, whether to include the original outlier SNP. Default is FALSE.}
}
\value{
List
}
\description{
This functions takes the output from the outlierAssociations function and
produces an instrument for each of the significant candidate traits previously
identified. The resulting SNP list can be exported using the exportSNPs function
to calculate associations with exposures and/or outcomes if these have been
calculated from individual-level data.
}
