% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateInstruments.R
\name{exportSNPs}
\alias{exportSNPs}
\title{Export candidate SNP list}
\usage{
exportSNPs(dat, filename)
}
\arguments{
\item{dat}{output from candidateInstruments function}

\item{filename}{filename for SNP list}
}
\value{

}
\description{
Exports list of SNPs which form part of the instrument for the candidate traits
identified. Each line represents one SNP. This file can be read by programs such
as bgenix (in BASH) which aid in SNP extraction from larger genomic datasets.
}
