% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{importPlink2}
\alias{importPlink2}
\title{Import Plink 2 Associations}
\usage{
importPlink2(association_file, variable_type, eaf_file, phenotype)
}
\arguments{
\item{association_file}{output from plink2 -glm function}

\item{variable_type}{choose from c("exposure", "outcome")}

\item{eaf_file}{.afreq output from plink2 detailing allele frequencies (optional)}

\item{phenotype}{string detailing variable of interest (optional)}
}
\value{
data frame
}
\description{
This function takes the output from the plink 2 -glm output and returns a
data frame readable by the TwoSampleMR package. Users can choose to include
.freq files with details concerning allele frequencies and these will be
included in the output data frame which aids in the process of harmonisation
when using the harmonise_data() function from the TwoSampleMR package.
}
