% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCandidateMR.R
\name{runCandidateMR}
\alias{runCandidateMR}
\title{Run MRTryx Candidate MR analyses}
\usage{
runCandidateMR(
  dat,
  candidate_outcome_associations,
  candidate_exposure_associations,
  mr_method = "mr_ivw",
  include_outliars = FALSE
)
}
\arguments{
\item{dat}{output from the CandidateInstruments function}

\item{candidate_outcome_associations}{formated data (formated using TwoSampleMR::format_data() or importPlink2()) for candidate SNP - outcome associations}

\item{candidate_exposure_associations}{formated data (formated using TwoSampleMR::format_data() or importPlink2()) for candidate SNP - exposure associations}

\item{mr_method}{default is IVW}

\item{include_outliars}{}
}
\value{
List
}
\description{
This function takes the output from the candidateInstruments() function
and formated data for the candidate-outcome associations and candidate-exposure
associations. These formated datasets are produced from the TwoSampleMR::format_data()
or the importPlink2() function.
}
