% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryxAdjustmentMV.R
\name{tryxAdjustmentMV}
\alias{tryxAdjustmentMV}
\title{MR-TRYX Multivariate Adjustment}
\usage{
tryxAdjustmentMV(tryxscan, lasso = TRUE, id_remove = NULL, proxies = FALSE)
}
\arguments{
\item{tryxscan}{Result from trys::sig() function}

\item{lasso}{Whether to shrink the estimates of each trait within SNP. Default=TRUE.}

\item{id_remove}{List of IDs to exclude from the adjustment analysis. It is possible that in the outlier search a candidate trait will come up which is essentially just a surrogate for the outcome trait (e.g. if you are analysing coronary heart disease as the outcome then a variable related to heart disease medication might come up as a candidate trait). Adjusting for a trait which is essentially the same as the outcome will erroneously nullify the result, so visually inspect the candidate trait list and remove those that are inappropriate.}

\item{proxies}{Look for proxies in the MVMR methods. Default = FALSE.}
}
\value{
List
}
\description{
This function takes the output from the tryx::sig() function and runs multivariateMR to
estimate adjusted candidate outlier - outcome associations.
}
